Refactor PnP validation components to be shared
