+++ src/components/PnpValidation/PnpValidationIcon.tsx
@@ -2,9 +2,9 @@ import { Check, Error, Warning } from '@mui/icons-material';
 import { Badge, IconButton, IconButtonProps, Tooltip } from '@mui/material';
 import { groupToMapBy } from '@seedcompany/common';
 import { PnpProblemSeverity as Severity } from '~/api/schema.graphql';
-import { useDialog } from '../../../components/Dialog';
-import { FileNodeInfoFragment } from '../../../components/files';
-import { useNumberFormatter } from '../../../components/Formatters';
+import { useDialog } from '../Dialog';
+import { FileNodeInfoFragment } from '../files';
+import { useNumberFormatter } from '../Formatters';
 import { PnPExtractionProblems } from './PnPExtractionProblems';
 import { PnpExtractionResultFragment } from './pnpExtractionResult.graphql';
 import { PnPExtractionResultDialog } from './PnpExtractionResultDialog';
