+++ src/scenes/ProgressReports/Detail/ProgressReportCard.tsx
@@ -1,24 +1,14 @@
 import { useMutation } from '@apollo/client';
-import { Preview as PreviewIcon } from '@mui/icons-material';
-import {
-  CircularProgress,
-  IconButton,
-  Stack,
-  Tooltip,
-  Typography,
-} from '@mui/material';
+import { CircularProgress, Stack, Typography } from '@mui/material';
 import {
   DefinedFileCard,
   DefinedFileCardProps,
 } from '../../../components/DefinedFileCard';
-import {
-  NonDirectoryActionItem as File,
-  FileActionsContextProvider,
-  useFileActions,
-} from '../../../components/files/FileActions';
+import { FileActionsContextProvider } from '../../../components/files/FileActions';
+import { PreviewIconButton } from '../../../components/files/FileActions/PreviewIconButton';
 import { UploadPeriodicReportFileDocument } from '../../../components/PeriodicReports/Upload/UpdatePeriodicReport.graphql';
-import { PnPReextractIconButton } from '../PnpValidation/PnPReextractIconButton';
-import { PnPValidationIcon } from '../PnpValidation/PnpValidationIcon';
+import { PnPReextractIconButton } from '../../../components/PnpValidation/PnPReextractIconButton';
+import { PnPValidationIcon } from '../../../components/PnpValidation/PnpValidationIcon';
 import { ReextractPnpProgressDocument } from '../PnpValidation/ReextractProgress.graphql';
 import { ProgressReportDetailFragment } from './ProgressReportDetail.graphql';
 
@@ -78,7 +68,7 @@ export const ProgressReportCard = ({ progressReport, ...rest }: Props) => {
               <Typography variant="h3">PnP File</Typography>
               {file && (
                 <>
-                  <Preview file={file} />
+                  <PreviewIconButton file={file} />
                   {reextracting ? (
                     <CircularProgress size={15} sx={{ ml: 1.1 }} />
                   ) : (
@@ -108,16 +98,5 @@ export const ProgressReportCard = ({ progressReport, ...rest }: Props) => {
         />
       </>
     </FileActionsContextProvider>
-  );
-};
-
-const Preview = ({ file }: { file: File }) => {
-  const { openFilePreview } = useFileActions();
-  return (
-    <Tooltip title="Preview">
-      <IconButton onClick={() => openFilePreview(file)} size="small">
-        <PreviewIcon />
-      </IconButton>
-    </Tooltip>
   );
 };
